# Conflicts:
#	shared/src/main/scala/com/neo/sk/tank/shared/ptcl/model/package.scala

+++ backend/src/main/scala/com/neo/sk/tank/core/tank/BulletServerImpl.scala
@@ -7,8 +7,8 @@ import com.neo.sk.tank.shared.ptcl.tank.Bullet
   * Created by hongruying on 2018/7/10
   */
 class BulletServerImpl (
-                         override val bId: Long,
-                         override val tankId: Long,
+                         override val bId: Int,
+                         override val tankId: Int,
                          override protected val startPosition: model.Point,
                          override protected val createTime: Long,
                          override val damage: Int,
